commit auditFunc

+++ src/main/java/com/lclgl/controller/FileController.java
@@ -1,16 +1,18 @@
 package com.lclgl.controller;
 
 import com.lclgl.service.FileService;
+import com.lclgl.service.ProService;
+import com.lclgl.service.StaffService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import java.io.File;
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 @RestController
@@ -19,29 +21,82 @@ public class FileController {
 
     @Autowired
     private FileService fileService;
+    @Autowired
+    private StaffService staffService;
+    @Autowired
+    private ProService proService;
 
     @PostMapping("/getFileList")
-    public Map<String, Object> fileList(String path) throws IOException {
+    public Map<String, Object> fileList(String path, HttpSession session) throws IOException {
+        if (session.getAttribute("staffId") == null) {
+            HashMap<String, Object> map = new HashMap<>();
+            map.put("msg", "你没有权限！");
+            return map;
+        }
         return fileService.fileList(path);
     }
 
-    @GetMapping("/downLoad")
-    public void downLoad(String fileName, String filePath, HttpServletResponse response) throws UnsupportedEncodingException {
-        fileService.downLoad(fileName, filePath, response);
+    @GetMapping("/download")
+    public void download(String fileName, String filePath, HttpServletResponse response, HttpSession session) throws UnsupportedEncodingException {
+        if (session.getAttribute("staffId") == null) {
+            HashMap<String, Object> map = new HashMap<>();
+            map.put("msg", "你没有权限！");
+            return;
+        }
+        fileService.download(fileName, filePath, response);
     }
 
-    @PostMapping("/upload")
-    public Map<String, Object> upload(MultipartFile file) {
-        return fileService.upload(file, new File("待审核文件/xxx宾馆设计/模型/").getAbsolutePath());
+    @GetMapping("/download/{fileName}/{auditStatus}/{staffId}/{proId}")
+    public void download(@PathVariable("fileName") String fileName,
+                         @PathVariable("auditStatus") String auditStatus,
+                         @PathVariable("staffId") String staffId,
+                         @PathVariable("proId") String proId,
+                         HttpServletResponse response,
+                         HttpSession session) throws UnsupportedEncodingException {
+        if (session.getAttribute("staffId") == null) {
+            HashMap<String, Object> map = new HashMap<>();
+            map.put("msg", "你没有权限！");
+            return;
+        }
+        String path = null;
+        String staffType = staffService.getStaffType(Integer.parseInt(staffId));
+        String proName = proService.getProName(Integer.parseInt(proId));
+        if ("审核通过".equals(auditStatus)) path = "项目列表/" + proName + "/" + staffType + "/";
+        else if ("待审核".equals(auditStatus) || "审核中".equals(auditStatus)) path = "待审核文件/" + proName + "/" + staffType + "/";
+        else path = "回收站/" + proName + "/" + staffType + "/";
+        path += fileName;
+        fileService.download(fileName, path, response);
+    }
+
+    @PostMapping("/upload/{commitWay}/{proId}")
+    public Map<String, Object> upload(MultipartFile file, @PathVariable("commitWay") String commitWay, @PathVariable("proId") String proId, HttpSession session) {
+        Object staffId = session.getAttribute("staffId");
+        if (staffId == null) {
+            HashMap<String, Object> map = new HashMap<>();
+            map.put("msg", "你没有权限！");
+            return map;
+        }
+        String path = fileService.getSavePath(commitWay, Integer.parseInt(proId), (int) staffId);
+        return fileService.upload(file, path, commitWay, Integer.parseInt(proId), (int) staffId);
     }
 
     @PostMapping("/delFile")
-    public Map<String, Object> delFile(String path) {
+    public Map<String, Object> delFile(String path, HttpSession session) {
+        if (session.getAttribute("staffId") == null) {
+            HashMap<String, Object> map = new HashMap<>();
+            map.put("msg", "你没有权限！");
+            return map;
+        }
         return fileService.delete("项目列表\\" + path);
     }
 
     @PostMapping("/mkpro/{proName}")
-    public Map<String, Object> mkpro(MultipartFile file, @PathVariable("proName") String proName) {
+    public Map<String, Object> mkpro(MultipartFile file, @PathVariable("proName") String proName, HttpSession session) {
+        if (session.getAttribute("staffId") == null) {
+            HashMap<String, Object> map = new HashMap<>();
+            map.put("msg", "你没有权限！");
+            return map;
+        }
         return fileService.mkpro(file, new File("项目列表/" + proName + "/").getAbsolutePath());
     }
